+++ src/MicroElements.Metadata.Xml/Xml/IXmlParserContext.cs
@@ -3,6 +3,7 @@
 
 using System.Collections.Concurrent;
 using MicroElements.Functional;
+using MicroElements.Metadata.Schema;
 
 namespace MicroElements.Metadata.Xml
 {
@@ -12,6 +13,11 @@ namespace MicroElements.Metadata.Xml
     public interface IXmlParserContext
     {
         /// <summary>
+        /// Gets <see cref="IXmlParserSettings"/> for this context.
+        /// </summary>
+        IXmlParserSettings ParserSettings { get; }
+
+        /// <summary>
         /// Gets messages list.
         /// </summary>
         IMutableMessageList<Message> Messages { get; }
@@ -20,6 +26,11 @@ namespace MicroElements.Metadata.Xml
         /// Gets parsers cache.
         /// </summary>
         ConcurrentDictionary<IProperty, IValueParser> ParsersCache { get; }
+
+        /// <summary>
+        /// Gets schema cache.
+        /// </summary>
+        ConcurrentDictionary<IProperty, ISchema> SchemaCache { get; }
     }
 
     /// <summary>
@@ -28,17 +39,36 @@ namespace MicroElements.Metadata.Xml
     public class XmlParserContext : IXmlParserContext
     {
         /// <inheritdoc />
+        public IXmlParserSettings ParserSettings { get; }
+
+        /// <inheritdoc />
         public IMutableMessageList<Message> Messages { get; }
 
         /// <inheritdoc />
         public ConcurrentDictionary<IProperty, IValueParser> ParsersCache { get; }
 
+        /// <inheritdoc />
+        public ConcurrentDictionary<IProperty, ISchema> SchemaCache { get; }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="XmlParserContext"/> class.
+        /// </summary>
+        /// <param name="parserSettings">Parser settings.</param>
+        /// <param name="messages">Optional message list.</param>
+        /// <param name="parsersCache">Optional parsers cache.</param>
+        /// <param name="schemaCache">Optional schemas cache.</param>
         public XmlParserContext(
+            IXmlParserSettings parserSettings,
             IMutableMessageList<Message>? messages = null,
-            ConcurrentDictionary<IProperty, IValueParser>? parsersCache = null)
+            ConcurrentDictionary<IProperty, IValueParser>? parsersCache = null,
+            ConcurrentDictionary<IProperty, ISchema>? schemaCache = null)
         {
+            parserSettings.AssertArgumentNotNull(nameof(parserSettings));
+
+            ParserSettings = parserSettings;
             Messages = messages ?? new MutableMessageList<Message>();
-            ParsersCache = parsersCache ?? new ConcurrentDictionary<IProperty, IValueParser>(comparer: PropertyComparer.ByReferenceComparer);
+            ParsersCache = parsersCache ?? new ConcurrentDictionary<IProperty, IValueParser>(comparer: parserSettings.PropertyComparer);
+            SchemaCache = schemaCache ?? new ConcurrentDictionary<IProperty, ISchema>(comparer: parserSettings.PropertyComparer);
         }
     }
 }